chore: Update dependencies and styles
This commit updates the dependencies in the `package.json` file, including Next.js to version 14.2.4, React to version 18, and adds `react-slick` and `slick-carousel` as new dependencies. It also updates the `tailwind.config.ts` file to include new colors for primary, secondary, and background. Additionally, it modifies the `globals.css` file to set the background color and text color using the new color variables.